* master:
  Added check of existence when delete resource
  tizenrt: Select LWIP if config enabled
  tizenrt: Defines OC_MAX_SESSION_EVENT_CBS
  tizenrt: Defines __TIZENRT__ for porting
  Fix integer overrun
  Fix variable type mismatch & add parentheses.
  Fix Windows build
  Adding NULL check to prevent crash
  Add Android port
  Pointers checks
  Define OC_UUID_LEN (37)
  Add clock & storage TC in platformtest.
  Win_ip:send_msg:set srcaddr if not saved apriori
  oc_endpoint: fix condition for DNS lookup
  oc_discovery:set interface idx in client handler
  Warnnings finded when compile g++ with permissive
  Adding dns lookup in ipadapter

+++ messaging/coap/coap.c
@@ -1168,7 +1168,7 @@ coap_set_header_uri_path(void *packet, const char *path, int path_len)
   }
 
   coap_pkt->uri_path = path;
-  coap_pkt->uri_path_len = path_len;
+  coap_pkt->uri_path_len = (size_t)path_len;
 
   SET_OPTION(coap_pkt, COAP_OPTION_URI_PATH);
   return coap_pkt->uri_path_len;
